consolidated StudentCard CSS and Module into same folder
